todosapp: fix input selection issue on sync
Before this CL, if person A was editing a todo (or tag, or list
name) while person B commits an edit and the committed edit gets
synced, the sync would trigger a UI update, causing A's todo text
(or tag, or list name) to get selected - and A's next keystroke
would wipe any existing text.

NOTE(adam): I'm not 100% sure this CL fixes the issue, because
it's difficult or impossible to test with a single laptop. I'll
have to test it in the office with Bindu.

Change-Id: If473402d869c134a8a33b6036417305caa076114