TODOs: Add Instrumentation Test for MainActivity and DataList Unit Test
Added a new flavor called Mock that is meant to run this Instrumentation
By getting rid of a Firebase/Syncbase dependency, we can verify that
- the UI triggers the correct Persistence actions.
- calls to the Listener correctly manipulate the UI

In other words, it's just up to the Persistence implementations to get
the logic between the two pieces correct.

The Instrumentation Test uses Mockito and Espresso.

The DataList got a few unit tests to confirm that it works as expected.

Change-Id: I6e50883e60ee5854054819c47fe03bce4207ea1d