TODOs: Primitive sharing. Changes Spec and "Poll" Spec for Sharees
The first step of sharing is to enable the sharers to pick who they
want to share to and see who they've shared with.

This CL accomplishes both by polling the SyncgroupSpec for members.
It also has a primitive dialog that allows specification of a person
to share to.

Caveats:
- no validation of email addresses
- polling has to be sufficient since there's no watch for spec
- no instrumentation tests yet

Change-Id: I93dde43254aab6561d6b9b06c64238df01dbc5cf