Enable remote inspection of syncbase/firebase.
This commit hooks up the Vanadium RemoteInspectors to the syncbase
server and also provides a means to send instructions for
debugging/inspecting the persistence layer (syncbase or firebase)
via email.

Of course, the way firebase is setup in the app right now, there are no
credentials required as the data store is accessible to all.

The syncbase setup provides a token that gives access to the debug
services for 1 day.

Change-Id: Id43affa50bee6b4cf774ce228e67cafe27b3d411