TODOs: Try to join with exponential backoff
Previously, we would only attempt to join a syncgroup twice. Now,
we'll do it up to 5 times (though we can change that default). These
tries are done with exponential backoff, so theoretically there is
no need for a limit.

The reason this wasn't included before is that it seemed to trigger
sync breakages; however, those problems have been resolved now that
syncgroup ad failures don't break sync anymore.

Change-Id: I7ba05a178c243718a24b5c2920a9a510be48fe25