TODOs: Move Edit List and Mark All Done to the overflow
Edit List is no longer an action on the toolbar, it's in overflow.
Mark All Done is no longer possible to swipe. It's now in the overflow
of the single list view. We don't have mark all as undone.

(To delete, you can just go into edit list, so I didn't expose it.)

UI tests were updated to correspond, as were the Firebase and Mock build
flavors.

Change-Id: I6cd2767bae1941549ce3b738aee53479a1d19244