TODOs: Upon creating, enter the list
This seemed like the right thing to do. After creating a list, you
would probably edit it immediately, so that's what happens now.

In order for that metadata to arrive earlier, syncgroup creation was
deferred and can be done at any time. Since this means that the userdata
collection may get data earlier than expected, the number of allowed
join retries was bumped up to 10.

Change-Id: I657be219aeed88f419a5696f1887326f2a9e040c