todos: Ensure that local syncbase address is loopback addr.
Sometimes, (when i try to add a bt address to the listenspec of the
local server), this code tried to make a local rpc using a bt address
which doesn't work. This change ensures that we return a loopback
address when we get a chance.
This new API will fix this issue.

Change-Id: Ic98fb638e5d013752a6cd6bbf9d7ab2e5e12e9a8