TODOs: Pass Permissions and rename db
Two changes related to Ivan's CLs for enforcing blessings on creation
- db now uses the app blessing portion only
  It doesn't use v.io:a:xyz (or whatever that is)
- Every thing we create (server, database, collection, syncgroup)
  now pass permissions. Notably, they only pass permissions with the
  correct set of tags. It is an error to specify more than the ones
  I give.

Effects: Current users will appear to lose all their data, even though
it is still there. The reason is that we are now accessing a completely
different database. New users should be unaffected by people who are
"behind" in the app.

Change-Id: Ieb34bca45a7254cc16da7ef7be0f5e458d846ea7