TODOs: When inviting, give Admin on the Syncgroup
If you give admin on the syncgroup, then others are allowed to join
at you in the neighborhood.

Note: This is meant for the demo, where two same user devices need to
share a shared list between themselves while disconnected from the
original creator.

We have a filter on the creator's email address. We only show the Share
button and create invitation advertisements for the creator, even though
technically, other admins could do it. (In other words, this change only
affects which devices you can use to join the syncgroup.)

Change-Id: I4a0a50dc6f4ed378bfaca9bee9d134244e6736be