TODOs: HLAPI 0.1.7 upgrade (cloud) and discovery workaround
This upgrades TODOs to use the cloud changes from 0.1.7.

It also contains a bugfix for finding peers. Restarting an ad
doesn't seem to play well with our underlying implementation,
so I'm more careful to only stop if started and start if not
started. However, I have a feeling the real bug is deeper in
our implementation.

Change-Id: I3232cca474b9aca1e1bd9bbc917bd2224d252f13